Re #1882 (misc): Removed stripping of '[]' from pvalue header parameters.
The stripping of '[]' from header parameters causes issues if
something (like a port) occurrs after the final ']'.

'[2001:a::b]' will correctly parse to '2001:a::b'
'[2001:a::b]:8080' will correctly parse to '2001:a::b' but the scanner is left
with ':8080' and parsing stops with a syntax error.

Thanks to Anthony Messina and George Joseph for the patch.